feat(geocoder): implement parallel processing to leverage full QPS
Google API request rates are measured in a bucket model.
Each second, up to 50 requests are allowed (as of June 2016).
The interval at which the requests arrive within a bucket are irrelevant.
This commit implements parallel processing for geocoding requests.
It also optimises the rate limitation code to make best use of the allowed QPS.